Add enums for all IDs
This allows content to be created more easily without fully memorizing the IDs, without modifying the existing structure too much. I tried to make things work with the ids fully converted into enums but its clear that Typescript just does not like it when keys are not strings or numbers. I think records may work for this, but iterating over them seemed to be a pain, and this solution seemed good enough.